+++ src/Elders.Pandora/IConfigurationRepository.cs
@@ -1,8 +1,12 @@
-﻿namespace Elders.Pandora
+﻿using System.Collections.Generic;
+
+namespace Elders.Pandora
 {
     public interface IConfigurationRepository
     {
         string Get(string key);
         void Set(string key, string value);
+        void Delete(string key);
+        IEnumerable<DeployedSetting> GetAll();
     }
 }