+++ src/Elders.Pandora/WindowsEnvironmentVariables.cs
@@ -1,10 +1,17 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
+using System.Text.RegularExpressions;
 
 namespace Elders.Pandora
 {
     public class WindowsEnvironmentVariables : IConfigurationRepository
     {
+        public void Delete(string key)
+        {
+            throw new NotSupportedException($"This operation is not supported for {nameof(WindowsEnvironmentVariables)}");
+        }
+
         public string Get(string key)
         {
             if (string.IsNullOrEmpty(key)) throw new ArgumentException(nameof(key));
@@ -14,6 +21,28 @@ namespace Elders.Pandora
                 throw new KeyNotFoundException("Unable to find environment variable " + key);
 
             return setting;
+        }
+
+        public IEnumerable<DeployedSetting> GetAll()
+        {
+            var regex = new Regex(@"([^@]+)@@([^\^]+)\^([^~]+)~~(.+)");
+
+            var all = Environment.GetEnvironmentVariables(EnvironmentVariableTarget.Machine);
+
+            foreach (DictionaryEntry item in all)
+            {
+                var result = regex.Match(item.Key.ToString());
+                if (result.Success)
+                {
+                    yield return new DeployedSetting(
+                        raw: result.Groups[0].Value,
+                        applicationName: result.Groups[1].Value,
+                        cluster: result.Groups[2].Value,
+                        machine: result.Groups[3].Value,
+                        key: result.Groups[4].Value,
+                        value: Environment.GetEnvironmentVariable(result.Groups[0].Value, EnvironmentVariableTarget.Machine));
+                }
+            }
         }
 
         public void Set(string key, string value)